+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * [MDCButton] Refactored example to be compatibile with catalog by convention [Randall Li](https://github.com/randallli)
 * Catalog by convention: Support duplicate hierarchy entries. [Randall Li](https://github.com/randallli)
 * [MDCCatalog] Catalog by convention grabs storyboard resources. [Randall Li](https://github.com/randallli)
+* [MDCSwitch] Refactored example to be compatible with catalog by convention
+  [Randall Li](https://github.com/randallli)
 * [MDCSlider] Refactored example to be compatibile with catalog by convention [Randall Li](https://github.com/randallli)
 * [gh-pages] Minor tiding of the preview script for gh-pages. [Jeff Verkoeyen](https://github.com/jverkoey)
 * [Conventions] Moved all docs assets into a `docs/assets` directory per component by